Remove standardize.py
