--- tests/reference/test_logic_standardizer_inplace_program.txt
@@ -1,4 +1,0 @@
-A_4 = Aggregate(add, 0, Reorder(MapJoin(mul, Table(A, i, i_2), Table(A_2, i_2, i_3)), i, i_2, i_3), i_2)
-A_5 = Aggregate(overwrite, 0, Reorder(MapJoin(add, Table(A_4, i, i_3), MapJoin(add, Table(A_3, i, i_3), Table(A_2, i, i_3))), i, i_3), )
-A_3 = Reorder(Table(A_5, i, i_3), i, i_3)
-return A_3
